Rename db model module.
